move exchange benchmarks to runtime dir
