--- pallets/exchange/benchmarking/Cargo.toml
@@ -1,63 +1,0 @@
-[package]
-name = "pallet-exchange-benchmarking"
-version = "3.3.10"
-description = "Exchange Benchmarking Module"
-authors = ["GalacticCouncil"]
-edition = "2021"
-homepage = "https://github.com/galacticcouncil/basilisk-node"
-license = "Apache 2.0"
-repository = "https://github.com/galacticcouncil/basilisk-node"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[build-dependencies]
-substrate-wasm-builder = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17" }
-
-[dependencies]
-codec = { default-features = false, features = ["derive"], package = "parity-scale-codec", version = "2.3.1" }
-scale-info = { version = "1.0", default-features = false, features = ["derive"] }
-serde = { features = ["derive"], optional = true, version = "1.0.136" }
-
-# Local dependencies
-pallet-xyk = { path = "../../xyk", default-features = false }
-pallet-exchange = { path = "../../exchange", default-features = false }
-primitives = { path = "../../../primitives", default-features = false }
-
-# ORML dependencies
-orml-tokens = { git = "https://github.com/open-web3-stack/open-runtime-module-library", rev = "aac79b3b31953381669a2ffa9b3e9bfe48e87f38", default-features = false }
-orml-traits = { git = "https://github.com/open-web3-stack/open-runtime-module-library", rev = "aac79b3b31953381669a2ffa9b3e9bfe48e87f38", default-features = false }
-
-# HydraDX dependencies
-pallet-asset-registry = { git = "https://github.com/galacticcouncil//warehouse", rev = "e428fd81f84781c4508df72f095e7bed6bb58ea1", default-features = false }
-hydradx-traits = { git = "https://github.com/galacticcouncil/warehouse", rev = "e428fd81f84781c4508df72f095e7bed6bb58ea1", default-features = false }
-
-# Substrate dependencies
-frame-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-frame-system-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-
-[dev-dependencies]
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.17", default-features = false }
-
-[features]
-default = ["std"]
-std = [
-    "serde",
-    "codec/std",
-    "frame-support/std",
-    "frame-system/std",
-    "sp-std/std",
-    "sp-runtime/std",
-    "primitives/std",
-    "hydradx-traits/std",
-    "pallet-exchange/std",
-    "pallet-asset-registry/std",
-    "frame-benchmarking/std",
-    "scale-info/std",
-]
